Reorganized route creation
